# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Qualys DSPM API
  description: |
    Data Security Posture Management API for discovering, classifying, and protecting sensitive data across cloud environments.

    ## Authentication
    All API endpoints (except /auth/login) require a JWT token. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Getting Started
    1. Login with `POST /api/v1/auth/login` to get a JWT token
    2. Use the token to authenticate subsequent requests
    3. Add a cloud account with `POST /api/v1/accounts`
    4. Trigger a scan with `POST /api/v1/accounts/{accountID}/scan`
  version: 1.0.0
  contact:
    name: DSPM Support
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: API v1
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login
      description: Authenticate with username and password to receive JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                  example: admin123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout
      security: [BearerAuth: []]
      responses:
        '200':
          description: Logout successful
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      security: [BearerAuth: []]
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /accounts:
    get:
      tags: [Accounts]
      summary: List cloud accounts
      security: [BearerAuth: []]
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountListResponse'
    post:
      tags: [Accounts]
      summary: Add cloud account
      security: [BearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
  /accounts/{accountID}:
    get:
      tags: [Accounts]
      summary: Get account details
      security: [BearerAuth: []]
      parameters:
        - $ref: '#/components/parameters/AccountID'
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
    delete:
      tags: [Accounts]
      summary: Delete account
      security: [BearerAuth: []]
      parameters:
        - $ref: '#/components/parameters/AccountID'
      responses:
        '204':
          description: Account deleted
  /accounts/{accountID}/scan:
    post:
      tags: [Accounts, Scans]
      summary: Trigger scan for account
      description: Start a new scan job for the specified account
      security: [BearerAuth: []]
      parameters:
        - $ref: '#/components/parameters/AccountID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                scan_type:
                  type: string
                  enum: [FULL, QUICK, CLASSIFICATION]
                  default: FULL
                  description: Type of scan to perform
      responses:
        '202':
          description: Scan started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  job_id:
                    type: string
                    format: uuid
  /assets:
    get:
      tags: [Assets]
      summary: List data assets
      security: [BearerAuth: []]
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: resource_type
          in: query
          schema:
            type: string
            enum: [s3_bucket, rds_instance, dynamodb_table]
        - name: sensitivity
          in: query
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW, UNKNOWN]
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetListResponse'
  /assets/{assetID}:
    get:
      tags: [Assets]
      summary: Get asset details
      security: [BearerAuth: []]
      parameters:
        - $ref: '#/components/parameters/AssetID'
      responses:
        '200':
          description: Asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
  /assets/{assetID}/classifications:
    get:
      tags: [Assets, Classifications]
      summary: Get classifications for asset
      security: [BearerAuth: []]
      parameters:
        - $ref: '#/components/parameters/AssetID'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Asset classifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificationListResponse'
  /findings:
    get:
      tags: [Findings]
      summary: List security findings
      security: [BearerAuth: []]
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: severity
          in: query
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, suppressed]
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of findings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingListResponse'
  /findings/{findingID}:
    get:
      tags: [Findings]
      summary: Get finding details
      security: [BearerAuth: []]
      parameters:
        - name: findingID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Finding details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
  /classifications:
    get:
      tags: [Classifications]
      summary: List all classifications
      security: [BearerAuth: []]
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [PII, PHI, PCI, SECRETS]
        - name: sensitivity
          in: query
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW]
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of classifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificationListResponse'
  /scans:
    get:
      tags: [Scans]
      summary: List scan jobs
      security: [BearerAuth: []]
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
      responses:
        '200':
          description: List of scans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanListResponse'
  /scans/{scanID}:
    get:
      tags: [Scans]
      summary: Get scan details
      security: [BearerAuth: []]
      parameters:
        - name: scanID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Scan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
  /scans/{scanID}/cancel:
    post:
      tags: [Scans]
      summary: Cancel running scan
      security: [BearerAuth: []]
      parameters:
        - name: scanID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Scan cancelled
  /rules:
    get:
      tags: [Rules]
      summary: List classification rules
      security: [BearerAuth: []]
      responses:
        '200':
          description: List of rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleListResponse'
    post:
      tags: [Rules]
      summary: Create custom rule
      security: [BearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleRequest'
      responses:
        '201':
          description: Rule created
  /rules/templates:
    get:
      tags: [Rules]
      summary: Get rule templates
      security: [BearerAuth: []]
      responses:
        '200':
          description: Available rule templates
  /rules/test:
    post:
      tags: [Rules]
      summary: Test rule pattern
      security: [BearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pattern:
                  type: string
                test_data:
                  type: string
      responses:
        '200':
          description: Test results
  /remediation:
    get:
      tags: [Remediation]
      summary: List remediation actions
      security: [BearerAuth: []]
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, EXECUTING, COMPLETED, FAILED, REJECTED, ROLLED_BACK]
      responses:
        '200':
          description: List of remediation actions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemediationListResponse'
    post:
      tags: [Remediation]
      summary: Create remediation action
      security: [BearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRemediationRequest'
      responses:
        '201':
          description: Action created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemediationAction'
  /remediation/summary:
    get:
      tags: [Remediation]
      summary: Get remediation summary
      security: [BearerAuth: []]
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Remediation summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemediationSummary'
  /remediation/definitions:
    get:
      tags: [Remediation]
      summary: Get available action definitions
      security: [BearerAuth: []]
      responses:
        '200':
          description: Action definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActionDefinition'
  /remediation/playbooks:
    get:
      tags: [Remediation]
      summary: Get remediation playbooks
      security: [BearerAuth: []]
      responses:
        '200':
          description: Available playbooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playbook'
...
  /remediation/{actionID}/rollback:
    post:
      tags: [Remediation]
      summary: Rollback completed action
      security: [BearerAuth: []]
      parameters:
        - $ref: '#/components/parameters/ActionID'
      responses:
        '200':
          description: Action rolled back
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 1000
...
  schemas:
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            expires_at:
              type: string
              format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        role:
          type: string
...
security:
  - BearerAuth: []